+++ src/main/java/ti4/buttons/ButtonListener.java
@@ -2402,7 +2402,7 @@ public class ButtonListener extends ListenerAdapter {
                 ButtonHelper.deleteMessage(event);
             }
         } else if (buttonID.startsWith("getNewTransaction_")) {
-            ButtonHelper.getNewTransaction(game, player, buttonID, event);
+            TransactionHelper.getNewTransaction(game, player, buttonID, event);
         } else if (buttonID.startsWith("rejectOffer_")) {
             Player p1 = game.getPlayerFromColorOrFaction(buttonID.split("_")[1]);
             if (p1 != null) {
@@ -2436,15 +2436,15 @@ public class ButtonListener extends ListenerAdapter {
             }
             TransactionHelper.acceptTransactionOffer(p1, player, game, event);
             ButtonHelper.deleteMessage(event);
-        } else if (buttonID.startsWith("sendOffer_")) {
-            ButtonHelper.sendOffer(game, player, buttonID, event);
+        // } else if (buttonID.startsWith("sendOffer_")) {
+        //     ButtonHelper.sendOffer(game, player, buttonID, event);
         } else if (buttonID.startsWith("offerToTransact_")) {
-            ButtonHelper.resolveOfferToTransact(game, player, buttonID, event);
+            TransactionHelper.resolveOfferToTransact(game, player, buttonID, event);
         } else if (buttonID.startsWith("newTransact_")) {
-            ButtonHelper.resolveSpecificTransButtonsNew(game, player, buttonID, event);
+            TransactionHelper.resolveSpecificTransButtonsNew(game, player, buttonID, event);
             ButtonHelper.deleteMessage(event);
         } else if (buttonID.startsWith("transact_")) {
-            ButtonHelper.resolveSpecificTransButtonsOld(game, player, buttonID, event);
+            TransactionHelper.resolveSpecificTransButtonsOld(game, player, buttonID, event);
         } else if (buttonID.startsWith("play_after_")) {
             String riderName = buttonID.replace("play_after_", "");
             ButtonHelper.addReaction(event, true, true, "Playing " + riderName, riderName + " Played");
@@ -2655,7 +2655,7 @@ public class ButtonListener extends ListenerAdapter {
             }
 
         } else if (buttonID.startsWith("send_")) {
-            ButtonHelper.resolveSpecificTransButtonPress(game, player, buttonID, event, true);
+            TransactionHelper.resolveSpecificTransButtonPress(game, player, buttonID, event, true);
             ButtonHelper.deleteMessage(event);
         } else if (buttonID.startsWith("replacePDSWithFS_")) {
             ButtonHelperFactionSpecific.replacePDSWithFS(buttonID, event, game, player, ident);
